IEBH-449: Move FileExplorer into fixture
