+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "e2e-tests"
-version = "0.1.8"
+version = "0.1.9"
 description = "Pilot-HDC end-to-end tests."
 readme = "README.md"
 requires-python = ">=3.13,<3.14"
